Add class for GUI
