+++ src/main/java/ui/Launcher.java
@@ -9,6 +9,6 @@ import duke.Duke;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Duke.class, args);
+        Application.launch(DukeInterface.class, args);
     }
 }
